refactor: 공통으로 사용되는 constant를 util class로 분리
