+++ src/main/java/wooteco/subway/domain/strategy/fare/age/FareDiscountAgeConstant.java
@@ -1,0 +1,6 @@
+package wooteco.subway.domain.strategy.fare.age;
+
+public class FareDiscountAgeConstant {
+
+    protected static final int DEDUCT_FARE = 350;
+}
