+++ src/main/java/wooteco/subway/domain/strategy/fare/distance/FareDistanceConstant.java
@@ -1,0 +1,11 @@
+package wooteco.subway.domain.strategy.fare.distance;
+
+public class FareDistanceConstant {
+
+    protected static final int BASIC_FARE = 1250;
+    protected static final int ADDITIONAL_FARE = 100;
+    protected static final int BASIC_DISTANCE = 10;
+    protected static final int STEP_ONE_DISTANCE = 50;
+    protected static final double STEP_ONE_CHARGE_DISTANCE = 5.0;
+    protected static final double STEP_TWO_CHARGE_DISTANCE = 8.0;
+}
